+++ Assets/Scripts/Map.cs
@@ -177,11 +177,11 @@ public class MapData
         for (int i = 0; i < pointsHorizontal; i++)
         {
             //X Grid placement  
-            double x = pointsHorizontalSeparation * (i + 0.5f);
+            float x = pointsHorizontalSeparation * (i + 0.5f);
             for (int j = 0; j < pointsVertical; j++)
             {
                 //Y Grid placement                
-                double y = pointsVerticalSeparation * (j + 0.5f);
+                float y = pointsVerticalSeparation * (j + 0.5f);
 
                 //Randomize angular offset
                 float angle = UnityEngine.Random.Range(0, Utilities.PI2);
@@ -191,8 +191,8 @@ public class MapData
                 float aa = a * a;
                 float bb = b * b;
                 float tanAngle = Mathf.Tan(angle);
-                double offX = ab / Mathf.Sqrt((bb) + (aa) * (tanAngle * tanAngle));
-                double offY = ab / Mathf.Sqrt((aa) + (bb) / (tanAngle * tanAngle));
+                float offX = ab / Mathf.Sqrt((bb) + (aa) * (tanAngle * tanAngle));
+                float offY = ab / Mathf.Sqrt((aa) + (bb) / (tanAngle * tanAngle));
 
                 //Quadrant check
                 if (angle > -Utilities.HALFPI && angle < Utilities.HALFPI)
